rsx: Move draw call related functions to their own class
