+++ rpcs3/Emu/RSX/GL/GLVertexBuffers.cpp
@@ -153,7 +153,7 @@ gl::vertex_upload_info GLGSRender::set_vertex_buffer()
 	m_profiler.start();
 
 	//Write index buffers and count verts
-	auto result = std::visit(draw_command_visitor(*m_index_ring_buffer, m_vertex_layout), get_draw_command(rsx::method_registers));
+	auto result = std::visit(draw_command_visitor(*m_index_ring_buffer, m_vertex_layout), m_draw_processor.get_draw_command(rsx::method_registers));
 
 	const u32 vertex_count = (result.max_index - result.min_index) + 1;
 	u32 vertex_base = result.min_index;
@@ -250,7 +250,7 @@ gl::vertex_upload_info GLGSRender::set_vertex_buffer()
 	}
 
 	//Write all the data
-	write_vertex_data_to_memory(m_vertex_layout, vertex_base, vertex_count, persistent_mapping.first, volatile_mapping.first);
+	m_draw_processor.write_vertex_data_to_memory(m_vertex_layout, vertex_base, vertex_count, persistent_mapping.first, volatile_mapping.first);
 
 	m_frame_stats.vertex_upload_time += m_profiler.duration();
 	return upload_info;
