+++ rpcs3/Emu/RSX/VK/VKDraw.cpp
@@ -730,7 +730,7 @@ void VKGSRender::emit_geometry(u32 sub_index)
 
 	if (state_flags & rsx::vertex_arrays_changed)
 	{
-		analyse_inputs_interleaved(m_vertex_layout);
+		m_draw_processor.analyse_inputs_interleaved(m_vertex_layout, current_vp_metadata);
 	}
 	else if (state_flags & rsx::vertex_base_changed)
 	{
